+++ web/modules/custom/qs_activity/src/Plugin/Block/EventsCollectionBlock.php
@@ -106,13 +106,16 @@ class EventsCollectionBlock extends BlockBase implements ContainerFactoryPluginI
         $variables['badges']['subscriptions']['pendings'] = $this->badgeManager->getSubscription($variables['events'], NULL);
 
         // From list of Events where current user has confirmed subscription.
-        $variables['badges']['subscriptions']['confirmed'] = $this->badgeManager->getSubscription($variables['events'], 1);
+        $variables['badges']['subscriptions']['confirmed'] = $this->badgeManager->getSubscription($variables['events'], TRUE);
 
-        // From list of Events number of pending subscriptions.
-        $variables['badges']['admin']['subscriptions']['pendings'] = [];
+        // From list of Activities get user privileges.
+        $variables['badges']['privileges'] = $this->badgeManager->getPrivileges([$variables['activity']]);
 
-        // From list of Events number of subscriptions.
-        $variables['badges']['admin']['subscriptions']['confirmed'] = [];
+        // From list of Events count pending subscriptions by given events.
+        $variables['badges']['subscriptions']['pendings_guests'] = $this->badgeManager->countSubscriptions($variables['events'], NULL);
+
+        // From list of Events count confirmed subscriptions by given events.
+        $variables['badges']['subscriptions']['confirmed_guests'] = $this->badgeManager->countSubscriptions($variables['events'], TRUE);
       }
     }