+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## NEXT RELEASE
+ - add Privilege's Badge on Activity by theme #318,  #315 & #316
+ - add Subscription's Badge, using highest Privilege color, on Activity by date #308 & #309
+ - add Subscription's Badge, using highest Privilege color, on My Activities #304
+ - update Subscription's Badge, using highest Privilege color, on Activity page #543
+ - update Subscription's Badge, using highest Privilege color, on Calendar Weekly/Monthly Cards - below Calendar Dots - #387
+ - add Privilege's Badge on My Photos #412
+ - add Privilege's Badge on Photos by theme #441, #442 & #443
+ - add 'Direct Subscription' whitout needing request for Organizer of activity #313
+ - add subscription to the author Maintainer of Activity when creating an Event #311
+ - add collapsable event card on 'My subscriptions' page
+ - add subscription textual information on 'My Subscription' card event body
+ - add Subscription's Badge, using Count of Guests Pending & Confirmed on events pill #310, #311, #312 & #313
+ - add Subscription's Badge, using Count of Guests Pending on Activitiy Teaser Card #317 & #319
+ - switch the Event Dashboard button in Activity detail page when Guests Pending
+ - add Subscription's Badge, using highest Privilege color,  on Calendar Weekly/Monthly Dots #386
+ - update Subscripton's Badge, using hiest Privilege color, on Activity page #309 & #309
+
+## NEXT RELEASE
+ - add pagination #585
+
 ## 1.0.1 (2018-03-01)
  - fix images upload on past events - #576
  - update README.md badges
@@ -58,4 +79,3 @@
 
 ## 0.0.0 (2017-07-14)
  - init empty repo
-